Coding Standards: Correct `db_version` check in `wp-admin/admin.php`.
Follow-up to [47785].

See #49239.
Built from https://develop.svn.wordpress.org/trunk@47787


git-svn-id: http://core.svn.wordpress.org/trunk@47563 1a063a9b-81f0-0310-95a4-ce76da25c4cd

+++ wp-admin/admin.php
@@ -36,6 +36,7 @@ require_once dirname( __DIR__ ) . '/wp-load.php';
 nocache_headers();
 
 if ( get_option( 'db_upgraded' ) ) {
+
 	flush_rewrite_rules();
 	update_option( 'db_upgraded', false );
 
@@ -45,7 +46,9 @@ if ( get_option( 'db_upgraded' ) ) {
 	 * @since 2.8.0
 	 */
 	do_action( 'after_db_upgrade' );
-} elseif ( get_option( 'db_version' ) !== $wp_db_version && empty( $_POST ) ) {
+
+} elseif ( (int) get_option( 'db_version' ) !== $wp_db_version && empty( $_POST ) ) {
+
 	if ( ! is_multisite() ) {
 		wp_redirect( admin_url( 'upgrade.php?_wp_http_referer=' . urlencode( wp_unslash( $_SERVER['REQUEST_URI'] ) ) ) );
 		exit;
